requirements.txt
asgiref==3.5.2
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
cloudinary==1.29.0
crispy-bootstrap5==0.6
cryptography==37.0.2
defusedxml==0.7.1
dj-database-url==0.5.0
Django==4.0.5
django-allauth==0.51.0
django-crispy-forms==1.14.0
django-etc==1.3.1
django-heroku==0.3.1
django-hitcount==1.3.5
gunicorn==20.1.0
idna==3.3
oauthlib==3.2.0
Pillow==9.1.1
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.4.0
python3-openid==3.2.0
requests==2.28.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
whitenoise==6.2.0
